Updated deleting & reconnecting calendars API integration.
When connecting a calendar while there is a pre-existing connected calendar, the UI shows the newly connected calendar as the primary and gives option to remove it. Previously, it would still show the previous calendar as primary.

And when removing a calendar, the UI shows the POPUP_STAGES.CONNECT_CALENDAR modal view if there are no remaining connected calendars. But if there is, the UI shows the POPUP_STAGES.MANAGE_CALENDAR view. Previously, it would show the user POPUP_STAGES.MANAGE_AVAILABILITY view.